+++ base/changelog.md
@@ -12,3 +12,4 @@
 *  New storage plugin: json
 *  New storage plugin: fastbit_compression
 *  Added reconfiguration at runtime
+*  New intermediate plugin: geoip